@@ -59,14 +59,16 @@ public static class Global
 	public static int BWTBlockExtraSize => BWTBlockSize <= 0x8000 ? 2 : BWTBlockSize <= 0x800000 ? 3 : BWTBlockSize <= 0x80000000 ? 4 : BWTBlockSize <= 0x8000000000 ? 5 : BWTBlockSize <= 0x800000000000 ? 6 : BWTBlockSize <= 0x80000000000000 ? 7 : 8;
 	public const int WordsListActualParts = 3;
 	public static UsedMethods PresentMethods { get; set; } = UsedMethods.CS1 | UsedMethods.HF1 | UsedMethods.LZ1 | UsedMethods.CS2 | UsedMethods.LZ2;
-	public static string[][] SHETEndinds { get; } = new[] { new[] { "ото", "ост", "каз", "оло", "ово", "ако", "ени", "отор", "каза", "ста", "оро", "ест", "енн", "ала", "ств", "ольк", "аст", "нно", "елов", "огд", "тве", "ове", "рос", "овор", "льн", "ель", "тел", "век", "ова", "вал", "чен", "еск", "али", "ани", "осл", "тил", "ног", "стр", "асс", "епе", "сти", "иче", "ние", "тор", "амо", "вер", "рем", "оди", "тся", "оле", "епер", "ате", "ась", "аме", "осто", "олж", "нов", "лся", "ожн", "бра", "ума", "рил", "лас", "ольш", "ите", "ьно", "про", "изн", "рин", "рит", "оте", "ину", "ром", "няз", "ско", "ись", "ыва", "ват", "тво", "уде", "ров", "овер", "дел", "сег", "ван", "лос", "тно", "льно", "ико", "ься", "ави", "ила", "одо", "рис", "огл", "вор", "или", "ение", "лен", "нул", "рост", "ерв", "оче", "рес", "ило", "дру", "вид", "тог", "нны", "рот", "мен", "омн", "ния", "они", "ичег", "аск", "ерн", "ает", "ный", "лис", "етс", "овс", "ски", "олов", "рив", "тьс", "оне", "ибу", "ибуд", "ика", "усс", "опр", "одн", "ови", "род", "усск", "чно", "осп", "увст", "руго", "ете", "сил", "лек", "ком", "луч", "еско", "ник", "лед", "нос", "тар", "ене", "жен", "ным", "сто", "увс", "ива", "дела", "стро", "ево", "ача", "око", "олн", "нач", "аро", "ысл", "нае", "акон", "едн", "ече", "омо", "ара", "кой", "тан", "торо", "азв", "роф", "обо", "орош", "рас", "апр", "оря", "але", "ости", "ные", "ран", "авл", "рог", "ков", "дес", "иво", "опро", "тель", "або", "ина", "лав", "прос", "отр", "ться", "пят", "ное", "лась", "рист", "еме", "чес", "ход", "став", "аве", "оня", "ения", "очт", "ось", "рои", "рош", "ист", "енк", "олод", "оск", "ажд", "тви", "лекс", "лько", "ака", "озв", "авн", "едо", "лаг", "озн", "рал", "енкр", "спо", "еши", "ельн", "апи", "ооб", "мот", "твен", "олу", "енно", "рый", "иль", "вен", "рик", "оста", "озд", "очн", "менн", "рен", "вобо", "воб", "кол", "редс", "оспо", "лись", "луча", "енкроф", "шенн", "оним", "згля", "мотр", "ерев", "амет", "олож", "родо", "лаго", "иког", "стре", "стан", "ство", "роти", "олжн", "ался", "аско", "альн", "оказ", "лось", "част", "ейст", "ичес", "ейча", "олон", "омна", "тран", "илос", "аста", "венн", "ност", "орог", "абот", "евск", "ость", "сегд", "айре", "собе", "роси", "мест", "озна", "евин", "ился", "онеч", "ител", "азал", "стор", "вано", "вать", "естн", "емно", "андр", "ервы", "атьс", "руги", "енны", "вова", "роиз", "айрес", "льни", "енщи", "вите", "воег", "асск", "атер", "ыстр", "равд", "исты", "реме", "ичик", "азум", "очем", "азго", "нтер", "ечес", "овол", "ербе", "ильн", "виде", "аход", "проч", "егод", "ожал", "озмо", "ывал", "ткры", "ообщ", "енер", "рани", "олуч", "нжен", "ажет", "ател", "пиле", "есят", "одум", "елик", "рест", "твеч", "ческ", "нени", "прав", "ович", "ерво", "ерес", "азыв", "лавн", "нима", "озво", "ольн", "начи", "ходи", "алас", "евол", "ельз", "овал", "еспо", "ущес", "еств", "ердц", "брат", "лжал", "алос", "лишк", "режд", "екот", "осмо", "риня", "влен", "отча", "ризн", "пилет", "аког", "ание", "илас", "апис", "апад", "жасн", "рошл", "дног", "авер", "звес", "ерну", "етро", "рият", "ичиков", "ется", "одит", "ыход", "ерберт", "исто", "тнош", "еньг", "коль", "елиг", "аскольников", "апра", "орот", "одня", "нула", "алис", "рихо", "ачин", "льны", "кого", "поко", "итьс", "жден", "ьств", "ении", "луша", "ивал", "олно", "ажды", "обра", "обст", "олже", "ывае", "трел", "ания", "вали", "бъяс", "аетс", "инов", "вере", "юбов", "сяко", "пуст", "ично", "твер", "ален", "оман", "чита", "ннос", "рекр", "скри", "оско", "ыраж", "ронс", "едел", "елат", "исьм", "днак", "ианс", "редп", "есел", "овил", "лыша", "бщес", "ришл", "леду", "атал", "спом", "риве", "амеч", "апит", "нутр", "осте", "осси", "нием", "арак", "ловн", "ногд", "оциа", "епри", "мерт", "автр", "онят", "тара", "амог", "нако", "жида", "рави", "илип", "стью", "ричи", "вала", "ысок", "евуш", "пере", "арст", "льст", "олез", "одим", "елен", "вори", "одоб", "овар", "повн", "нить", "бращ", "траш", "редл", "альш", "осту", "есча", "остр", "снов", "остоевского", "рова", "дивл", "виже", "риго", "ораз", "риба", "еобх", "атали", "рого", "олст", "райн", "еком", "рево", "ледо", "ленн", "тави", "илис", "ован", "идим", "амер", "идет", "озвр", "варт", "евоч", "огож", "ереб", "одош", "асст", "аясь", "атур", "лыбк", "оэто", "тавл", "мысл", "аждо", "удьб", "естр", "лучи", "епре", "резв", "хват", "стве", "иозн", "тлич", "лени", "скор", "трад", "воре", "пред", "ухов", "авши", "спуг", "елан", "тпра", "дова", "ешит", "ства", "рохо", "ебед", "риех", "ержа", "лубо", "овто", "адос", "остоевский", "рудн", "ивши", "отер", "тить", "ычай", "роис", "есмо", "очув", "ерди", "держ", "ообр", "амен", "посо", "едео", "овна", "нный", "ероя", "асил", "йрто", "улся", "еобы", "ливо", "трас", "коре", "режн", "елал", "ошад", "риро", "озяй", "ульт", "асме", "яжел", "осуд", "змен", "евер", "омен", "озду", "рини", "аков", "овен", "роше", "нача", "ашег", "ными", "лыбн", "тано", "етыр", "озда", "роко", "умае", "вида", "обро", "икол", "нным", "сяки", "жени", "гром", "ится", "рово", "жнос", "анно", "ешен", "ласт", "ихин", "конч", "рове", "рика", "асте", "стат", "расн", "омне", "изав", "итат", "вает", "едеон", "вить", "акры", "исте", "оват", "ебен", "динс", "гнов", "елае", "ески", "ерье", "аспо", "оряд", "омощ", "анны", "окто", "ечен", "ойде", "дово", "олот", "льна", "риме", "диви", "орал", "тари", "овес", "чени", "спол", "ертв", "арен", "кнов", "ноги", "казы", "вадц", "ерем", "озра", "оссии", "вилс", "тказ", "нуть", "ромк", "аскр", "авал", "есьм", "стин", "знат", "ркад", "тнос", "ривы", "укам", "рони", "аним", "ругу", "ассу", "азва", "атил", "улас", "мени", "аклю", "олен", "ненн", "ложн", "одил", "ронский", "ести", "ужчи", "оряч", "дыва", "аете", "лиен", "альч", "охож", "твор", "рожа", "астасья", "наеш", "праш", "еуже", "лава", "рина", "олне", "вало", "азде", "отре", "адум", "ашем", "ассм", "бяза", "апро", "орил", "ению", "рипо", "тепа", "ский", "рикн", "окру", "анят", "одст", "вила", "ужик", "вяза", "ивот", "глая", "овре", "евоз", "стви", "оскл", "одоз", "аска", "бнос", "ятел", "астр", "ассе", "чите", "ончи", "фьев", "дцат", "асту", "авло", "бург", "твов", "иван", "ерны", "тдел", "тств", "луби", "удущ", "вопо", "юбоп", "рили", "нной", "дени", "реть", "ират", "лександрович", "данн", "слов", "ками", "азны", "лист", "вановна", "еспр", "отит", "ереж", "еетс", "етве", "итер", "есно", "едос", "одро", "ворц", "рени", "ерен", "офия", "кнул", "лексей", "адеж", "сход", "пасн", "лыба", "наль", "орфи", "быкн", "сили", "арти", "бычн", "слыш", "спыт", "спок", "полн", "трем", "омни", "олаг", "можн", "етрович", "аздр", "ослу", "еожи", "нног", "риде", "апри", "саен", "саенко", "тару", "рови", "асно", "апря", "бежд", "клон", "емей", "нные", "длеж", "едов", "ений", "аблю", "ласс", "одхо", "оспе", "ника", "звол", "ранц", "тчая", "обак", "обир", "счез", "иров", "ебол", "осад", "ораб", "янно", "вили", "уман", "веря", "ванович", "окра", "ному", "озьм", "лавя", "слав", "алек", "емен", "асче", "ающи", "нулс", "ажен", "авли", "ротя", "йртон", "аешь", "доро", "лать", "озяи", "рать", "втор", "нное", "репк", "родн", "одни", "рине", "дино", "омог", "рише", "азда", "ебедев", "урно", "деал", "бсто", "езна", "линн", "ошло", "енав", "вген", "ивая", "еньш", "чеви", "рива", "нико", "апол", "ольз", "оход", "атор", "ятно", "елом", "орма", "исти", "омещ", "верд", "ывая", "ресл", "арма", "вроп", "щени", "тепе", "еизв", "сихо", "азумихин", "робо", "ойти", "звин", "ожде", "азре", "ожен", "мина", "ысле", "озбу", "олнц", "огат", "овет", "ечег", "склю", "усты", "айте", "ыват", "ледн", "орит", "логи", "урна", "авеч", "илипповна", "очеш", "ског", "арго", "нофи", "вшис", "иват", "ажно", "ереп", "тиче", "роце", "едор", "огожин", "раси", "кажи", "огов", "азви", "вляе", "аньш", "одру", "итал", "есси", "кова", "тепан", "ченн", "альц", "тный", "овин", "шего", "ываю", "рича", "дите", "исат", "ссле", "лександровна", "елич", "няги", "шись", "роща", "квоз", "азли", "влял", "печа", "уйст", "нная", "амил", "вани", "инко", "елле", "ирок", "окры", "икну", "пуск", "елез", "ству", "адно", "ракт", "изавета", "овод", "сюша", "ступ", "оздр", "умат", "рибл", "оцел" },
-		new[] { "а", "я", "ы", "и", "е", "у", "ю", "ой", "ей", "ам", "ям", "ами", "ями", "ах", "ях", "о", "ь", "ом", "ем", "ём", "ов", "ью", "ени", "енем", "ен", "ён", "ян", "енам", "енами", "енах" }, new[] { "ый", "ий", "ого", "его", "ому", "ему", "ым", "им", "ая", "яя", "ую", "юю", "ою", "ею", "ое", "ее", "ые", "ие", "ых", "их", "ной", "ный", "ний", "ного", "него", "ному", "нему", "ным", "ним", "ном", "нем", "ная", "няя", "ней", "ную", "нюю", "ною", "нею", "ное", "нее", "ные", "ние", "ных", "них", "он", "на", "ня", "но", "нё", "ны", "ни" }, new[] { "ть", "ать", "еть", "ить", "оть", "уть", "ыть", "ять", "овать", "евать", "л", "ал", "ел", "ил", "ол", "ул", "ыл", "ял", "овал", "евал", "ла", "ала", "ела", "ила", "ола", "ула", "ыла", "яла", "овала", "евала", "ло", "ало", "ело", "ило", "оло", "уло", "ыло", "яло", "овало", "евало", "ли", "али", "ели", "или", "оли", "ули", "ыли", "яли", "овали", "евали", "ишь", "ит", "ите", "ешь", "ёшь", "ет", "ёт", "ете", "ёте", "аешь", "аёшь", "ает", "аёт", "аем", "аём", "аете", "аёте", "еешь", "еет", "еем", "еете", "иешь", "иёшь", "иет", "иёт", "ием", "иём", "иете", "иёте", "оешь", "оет", "оем", "оете", "уешь", "уёшь", "ует", "уёт", "уем", "уём", "уете", "уёте", "ьешь", "ьёшь", "ьет", "ьёт", "ьем", "ьём", "ьете", "ьёте", "юешь", "юёшь", "юет", "юёт", "юем", "юём", "юете", "юёте", "яешь", "яет", "яем", "яете", "ут", "ют", "ают", "еют", "иют", "оют", "уют", "ьют", "юют", "яют", "ат", "ят", "ться", "аться", "еться", "иться", "оться", "уться", "ыться", "яться", "оваться", "еваться", "лся", "ался", "елся", "ился", "олся", "улся", "ылся", "ялся", "овался", "евался", "лась", "алась", "елась", "илась", "олась", "улась", "ылась", "ялась", "овалась", "евалась", "лось", "алось", "елось", "илось", "олось", "улось", "ылось", "ялось", "овалось", "евалось", "лись", "ались", "елись", "ились", "олись", "улись", "ылись", "ялись", "овались", "евались", "усь", "юсь", "ишься", "ится", "имся", "итесь", "ешься", "ёшься", "ется", "ётся", "емся", "ёмся", "етесь", "ётесь", "аешься", "аёшься", "ается", "аётся", "аемся", "аёмся", "аетесь", "аётесь", "еешься", "еется", "еемся", "еетесь", "иется", "иётся", "иемся", "иёмся", "иетесь", "иётесь", "оешься", "оется", "оемся", "оетесь", "уешься", "уёшься", "уется", "уётся", "уемся", "уёмся", "уетесь", "уётесь", "ьешься", "ьёшься", "ьется", "ьётся", "ьемся", "ьёмся", "ьетесь", "ьётесь", "юешься", "юёшься", "юется", "юётся", "юемся", "юёмся", "юетесь", "юётесь", "яешься", "яется", "яемся", "яетесь", "утся", "ются", "аются", "еются", "иются", "оются", "уются", "ьются", "юются", "яются", "атся", "ятся", "й", "ай", "ей", "ой", "уй", "ый", "яй", "йся", "айся", "ейся", "ойся", "уйся", "ыйся", "яйся", "ись", "ься", "ущий", "ющий", "ающий", "еющий", "иющий", "оющий", "ующий", "ьющий", "юющий", "яющий", "ащий", "ящий", "ущего", "ющего", "ающего", "еющего", "иющего", "оющего", "ующего", "ьющего", "юющего", "яющего", "ащего", "ящего", "ущему", "ющему", "ающему", "еющему", "иющему", "оющему", "ующему", "ьющему", "юющему", "яющему", "ащему", "ящему", "ущим", "ющим", "ающим", "еющим", "иющим", "оющим", "ующим", "ьющим", "юющим", "яющим", "ащим", "ящим", "ущем", "ющем", "ающем", "еющем", "иющем", "оющем", "ующем", "ьющем", "юющем", "яющем", "ащем", "ящем", "ущая", "ющая", "ающая", "еющая", "иющая", "оющая", "ующая", "ьющая", "юющая", "яющая", "ащая", "ящая", "ущей", "ющей", "ающей", "еющей", "иющей", "оющей", "ующей", "ьющей", "юющей", "яющей", "ащей", "ящей", "ущую", "ющую", "ающую", "еющую", "иющую", "оющую", "ующую", "ьющую", "юющую", "яющую", "ащую", "ящую", "ущее", "ющее", "ающее", "еющее", "иющее", "оющее", "ующее", "ьющее", "юющее", "яющее", "ащее", "ящее", "ущие", "ющие", "ающие", "еющие", "иющие", "оющие", "ующие", "ьющие", "юющие", "яющие", "ащие", "ящие", "ущих", "ющих", "ающих", "еющих", "иющих", "оющих", "ующих", "ьющих", "юющих", "яющих", "ащих", "ящих", "ущими", "ющими", "ающими", "еющими", "иющими", "оющими", "ующими", "ьющими", "юющими", "яющими", "ащими", "ящими", "ущийся", "ющийся", "ающийся", "еющийся", "иющийся", "оющийся", "ующийся", "ьющийся", "юющийся", "яющийся", "ащийся", "ящийся", "ущегося", "ющегося", "ающегося", "еющегося", "иющегося", "оющегося", "ующегося", "ьющегося", "юющегося", "яющегося", "ащегося", "ящегося", "ущемуся", "ющемуся", "ающемуся", "еющемуся", "иющемуся", "оющемуся", "ующемуся", "ьющемуся", "юющемуся", "яющемуся", "ащемуся", "ящемуся", "ущимся", "ющимся", "ающимся", "еющимся", "иющимся", "оющимся", "ующимся", "ьющимся", "юющимся", "яющимся", "ащимся", "ящимся", "ущемся", "ющемся", "ающемся", "еющемся", "иющемся", "оющемся", "ующемся", "ьющемся", "юющемся", "яющемся", "ащемся", "ящемся", "ущаяся", "ющаяся", "ающаяся", "еющаяся", "иющаяся", "оющаяся", "ующаяся", "ьющаяся", "юющаяся", "яющаяся", "ащаяся", "ящаяся", "ущейся", "ющейся", "ающейся", "еющейся", "иющейся", "оющейся", "ующейся", "ьющейся", "юющейся", "яющейся", "ащейся", "ящейся", "ущуюся", "ющуюся", "ающуюся", "еющуюся", "иющуюся", "оющуюся", "ующуюся", "ьющуюся", "юющуюся", "яющуюся", "ащуюся", "ящуюся", "ущееся", "ющееся", "ающееся", "еющееся", "иющееся", "оющееся", "ующееся", "ьющееся", "юющееся", "яющееся", "ащееся", "ящееся", "ущиеся", "ющиеся", "ающиеся", "еющиеся", "иющиеся", "оющиеся", "ующиеся", "ьющиеся", "юющиеся", "яющиеся", "ащиеся", "ящиеся", "ущихся", "ющихся", "ающихся", "еющихся", "иющихся", "оющихся", "ующихся", "ьющихся", "юющихся", "яющихся", "ащихся", "ящихся", "ущимися", "ющимися", "ающимися", "еющимися", "иющимися", "оющимися", "ующимися", "ьющимися", "юющимися", "яющимися", "ащимися", "ящимися", "вший", "авший", "евший", "ивший", "овший", "увший", "ывший", "явший", "овавший", "евавший", "вшего", "авшего", "евшего", "ившего", "овшего", "увшего", "ывшего", "явшего", "овавшего", "евавшего", "вшему", "авшему", "евшему", "ившему", "овшему", "увшему", "ывшему", "явшему", "овавшему", "евавшему", "вшим", "авшим", "евшим", "ившим", "овшим", "увшим", "ывшим", "явшим", "овавшим", "евавшим", "вшем", "авшем", "евшем", "ившем", "овшем", "увшем", "ывшем", "явшем", "овавшем", "евавшем", "вшая", "авшая", "евшая", "ившая", "овшая", "увшая", "ывшая", "явшая", "овавшая", "евавшая", "вшей", "авшей", "евшей", "ившей", "овшей", "увшей", "ывшей", "явшей", "овавшей", "евавшей", "вшую", "авшую", "евшую", "ившую", "овшую", "увшую", "ывшую", "явшую", "овавшую", "евавшую", "вшее", "авшее", "евшее", "ившее", "овшее", "увшее", "ывшее", "явшее", "овавшее", "евавшее", "вшие", "авшие", "евшие", "ившие", "овшие", "увшие", "ывшие", "явшие", "овавшие", "евавшие", "вших", "авших", "евших", "ивших", "овших", "увших", "ывших", "явших", "овавших", "евавших", "вшими", "авшими", "евшими", "ившими", "овшими", "увшими", "ывшими", "явшими", "овавшими", "евавшими", "вшийся", "авшийся", "евшийся", "ившийся", "овшийся", "увшийся", "ывшийся", "явшийся", "овавшийся", "евавшийся", "вшегося", "авшегося", "евшегося", "ившегося", "овшегося", "увшегося", "ывшегося", "явшегося", "овавшегося", "евавшегося", "вшемуся", "авшемуся", "евшемуся", "ившемуся", "овшемуся", "увшемуся", "ывшемуся", "явшемуся", "овавшемуся", "евавшемуся", "вшимся", "авшимся", "евшимся", "ившимся", "овшимся", "увшимся", "ывшимся", "явшимся", "овавшимся", "евавшимся", "вшемся", "авшемся", "евшемся", "ившемся", "овшемся", "увшемся", "ывшемся", "явшемся", "овавшемся", "евавшемся", "вшаяся", "авшаяся", "евшаяся", "ившаяся", "овшаяся", "увшаяся", "ывшаяся", "явшаяся", "овавшаяся", "евавшаяся", "вшейся", "авшейся", "евшейся", "ившейся", "овшейся", "увшейся", "ывшейся", "явшейся", "овавшейся", "евавшейся", "вшуюся", "авшуюся", "евшуюся", "ившуюся", "овшуюся", "увшуюся", "ывшуюся", "явшуюся", "овавшуюся", "евавшуюся", "вшееся", "авшееся", "евшееся", "ившееся", "овшееся", "увшееся", "ывшееся", "явшееся", "овавшееся", "евавшееся", "вшиеся", "авшиеся", "евшиеся", "ившиеся", "овшиеся", "увшиеся", "ывшиеся", "явшиеся", "овавшиеся", "евавшиеся", "вшихся", "авшихся", "евшихся", "ившихся", "овшихся", "увшихся", "ывшихся", "явшихся", "овавшихся", "евавшихся", "вшимися", "авшимися", "евшимися", "ившимися", "овшимися", "увшимися", "ывшимися", "явшимися", "овавшимися", "евавшимися",
-		"имый", "емый", "аемый", "еемый", "оемый", "уемый", "юемый", "яемый", "омый", "имого", "емого", "аемого", "еемого", "оемого", "уемого", "юемого", "яемого", "омого", "имому", "емому", "аемому", "еемому", "оемому", "уемому", "юемому", "яемому", "омому", "имым", "емым", "аемым", "еемым", "оемым", "уемым", "юемым", "яемым", "омым", "имом", "емом", "аемом", "еемом", "оемом", "уемом", "юемом", "яемом", "омом", "имая", "емая", "аемая", "еемая", "оемая", "уемая", "юемая", "яемая", "омая", "имой", "емой", "аемой", "еемой", "оемой", "уемой", "юемой", "яемой", "омой", "имую", "емую", "аемую", "еемую", "оемую", "уемую", "юемую", "яемую", "омую", "имое", "емое", "аемое", "еемое", "оемое", "уемое", "юемое", "яемое", "омое", "имые", "емые", "аемые", "еемые", "оемые", "уемые", "юемые", "яемые", "омые", "имых", "емых", "аемых", "еемых", "оемых", "уемых", "юемых", "яемых", "омых", "имыми", "емыми", "аемыми", "еемыми", "оемыми", "уемыми", "юемыми", "яемыми", "омыми", "аный", "еный", "ёный", "яный", "ованый", "еваный", "ёваный", "анный", "енный", "ённый", "янный", "ованный", "еванный", "ёванный", "аного", "еного", "ёного", "яного", "ованого", "еваного", "ёваного", "анного", "енного", "ённого", "янного", "ованного", "еванного", "ёванного", "аному", "еному", "ёному", "яному", "ованому", "еваному", "ёваному", "анному", "енному", "ённому", "янному", "ованному", "еванному", "ёванному", "аным", "еным", "ёным", "яным", "ованым", "еваным", "ёваным", "анным", "енным", "ённым", "янным", "ованным", "еванным", "ёванным", "аном", "еном", "ёном", "яном", "ованом", "еваном", "ёваном", "анном", "енном", "ённом", "янном", "ованном", "еванном", "ёванном", "аная", "еная", "ёная", "яная", "ованая", "еваная", "ёваная", "анная", "енная", "ённая", "янная", "ованная", "еванная", "ёванная", "аной", "еной", "ёной", "яной", "ованой", "еваной", "ёваной", "анной", "енной", "ённой", "янной", "ованной", "еванной", "ёванной", "аную", "еную", "ёную", "яную", "ованую", "еваную", "ёваную", "анную", "енную", "ённую", "янную", "ованную", "еванную", "ёванную", "аное", "еное", "ёное", "яное", "ованое", "еваное", "ёваное", "анное", "енное", "ённое", "янное", "ованное", "еванное", "ёванное", "аные", "еные", "ёные", "яные", "ованые", "еваные", "ёваные", "анные", "енные", "ённые", "янные", "ованные", "еванные", "ёванные", "аных", "еных", "ёных", "яных", "ованых", "еваных", "ёваных", "анных", "енных", "ённых", "янных", "ованных", "еванных", "ёванных", "аными", "еными", "ёными", "яными", "оваными", "еваными", "ёваными", "анными", "енными", "ёнными", "янными", "ованными", "еванными", "ёванными", "тый", "атый", "етый", "итый", "отый", "утый", "ытый", "ятый", "того", "атого", "етого", "итого", "отого", "утого", "ытого", "ятого", "тому", "атому", "етому", "итому", "отому", "утому", "ытому", "ятому", "тым", "атым", "етым", "итым", "отым", "утым", "ытым", "ятым", "том", "атом", "етом", "итом", "отом", "утом", "ытом", "ятом", "тая", "атая", "етая", "итая", "отая", "утая", "ытая", "ятая", "той", "атой", "етой", "итой", "отой", "утой", "ытой", "ятой", "тую", "атую", "етую", "итую", "отую", "утую", "ытую", "ятую", "тое", "атое", "етое", "итое", "отое", "утое", "ытое", "ятое", "тые", "атые", "етые", "итые", "отые", "утые", "ытые", "ятые", "тых", "атых", "етых", "итых", "отых", "утых", "ытых", "ятых", "тыми", "атыми", "етыми", "итыми", "отыми", "утыми", "ытыми", "ятыми", "учи", "ючи", "аючи", "еючи", "иючи", "оючи", "уючи", "ьючи", "юючи", "яючи", "в", "ав", "ев", "ив", "ув", "ыв", "яв", "овав", "евав", "вши", "авши", "евши", "ивши", "овши", "увши", "ывши", "явши", "овавши", "евавши", "ши", "ась", "ясь", "учись", "ючись", "аючись", "еючись", "иючись", "оючись", "уючись", "ьючись", "юючись", "яючись", "вшись", "авшись", "евшись", "ившись", "овшись", "увшись", "ывшись", "явшись", "овавшись", "евавшись", "шись" },
-		new[] { "без", "безо", "близ", "в", "вблизи", "ввиду", "вглубь", "вдогон", "вдоль", "взамен", "включая", "вкруг", "вместо", "вне", "внизу", "внутри", "внутрь", "во", "вовнутрь", "возле", "вокруг", "вопреки", "впереди", "вроде", "вслед", "вследствие", "встречу", "выключая", "для", "до", "за", "заместо", "из", "изнутри", "изо", "исключая", "к", "касаемо", "касательно", "ко", "кончая", "кроме", "кругом", "меж", "между", "мимо", "на", "наверху", "навстречу", "над", "надо", "назад", "назло", "накануне", "наперекор", "наперерез", "наподобие", "напротив", "насчёт", "ниже", "о", "об", "обо", "обок", "около", "от", "относительно", "ото", "перед", "передо", "по", "поверх", "под", "под видом", "подле", "подо", "подобно", "позади", "помимо", "поперёд", "поперёк", "порядка", "посередине", "после", "посреди", "посредине", "посредством", "пред", "предо", "прежде", "при", "про", "против", "путём", "ради", "с", "сверх", "сверху", "свыше", "сзади", "сквозь", "снизу", "со", "согласно", "спустя", "среди", "средь", "сродни", "супротив", "у", "через", "черезо", "чрез" },
-		new[] { "что", "как", "его", "это", "было", "все", "так", "она", "меня", "только", "мне", "был", "уже", "еще", "даже", "быть", "для", "ему", "сказал", "была", "когда", "очень", "они", "или", "чтобы", "есть", "него", "теперь", "себя", "если", "может", "чем", "были", "того", "время", "вас", "нибудь", "вдруг", "раз", "себе", "вот", "потому", "ничего", "этом", "этого", "нет", "человек", "вам", "будет", "сказала", "тебя", "этот", "где", "тем", "опять", "ведь", "под", "сам", "тебе", "без", "князь", "можно", "ней", "тоже", "который", "более", "ним", "тут", "всего", "больше", "человека", "всех", "жизни", "том", "мог", "нас", "потом", "там", "почти", "при", "совсем", "которые", "один", "своей", "дело", "них", "здесь", "сказать", "надо", "несколько", "эти", "тогда", "после", "свою", "глаза", "всегда", "сейчас", "какой", "кто", "день", "наконец", "знаю", "этой", "совершенно", "чего", "будто", "между", "два", "хотя", "эту", "такой", "никогда", "жизнь", "говорил", "через", "спросил", "ответил", "нам", "минуту", "нее", "хотел", "тот", "просто", "своего", "стал", "много", "над", "всем", "руку", "нужно", "другой", "слова", "руки", "перед", "хорошо", "про", "точно", "хоть", "мой", "лицо", "вместе", "именно", "людей", "свои", "эта", "лучше", "конечно", "такое", "нем", "кажется", "тому", "своим", "самом", "слишком", "которая", "тотчас", "нельзя", "три", "лет", "своих", "всю", "знал", "лишь", "чуть", "таки", "могу", "собой", "чтоб", "одно", "снова", "прежде", "этим", "голову", "которого", "нему", "одного", "стало", "весь", "свое", "почему", "давно", "свой", "которой", "хочу", "говорить", "должен", "знает", "пока", "прямо", "стороны", "деле", "мысли", "деньги", "особенно", "которых", "назад", "сама", "друг", "которое", "могла", "мысль", "люди", "этих", "мной", "моей", "довольно", "сделать", "впрочем", "одной", "колонисты", "заметил", "времени", "думал", "делать", "самого", "говорит", "одна", "быстро", "могли", "видел", "тех", "казалось", "дня", "действительно", "сегодня", "куда", "которую", "пор", "ответила", "случае", "против", "начал", "вопрос", "долго", "моя", "образом", "дверь", "оно", "какое", "ними", "сколько", "таким", "иногда", "своем", "дома", "первый", "спросила", "место", "какие", "кроме", "инженер", "часов", "немного", "буду", "знаете", "словно", "какая", "князя", "ночь", "отвечал", "такие", "свободы", "продолжал", "этому", "самое", "две", "мере", "мои", "сюда", "должно", "правда", "каким", "вся", "русской", "пять", "каждый", "сами", "слово", "часто", "острова", "сердце", "подумал", "тобой", "одну", "вообще", "однако", "двух", "любви", "хотела", "минут", "стоял", "такого", "комнату", "никто", "пред", "вами", "других", "затем", "должны", "домой", "кого", "час", "духа", "зачем", "ужасно", "говорю", "месте", "пошел", "самой", "генерал", "завтра", "вышел", "вчера", "глазами", "значит", "мало", "такая", "стала", "видеть", "голос", "должна", "дела", "говорила", "несмотря", "таких", "лица", "крайней", "жить", "весьма", "которым", "чувствовал", "хочет", "скоро", "понимаю", "года", "взгляд", "непременно", "котором", "знать", "взял", "хотелось", "имел", "сторону", "начала", "руками", "мое", "стали", "мир", "вперед", "мою", "остров", "места", "сразу", "сделал", "берегу", "равно", "видно", "женщина", "души", "самый", "друга", "своему", "смотрел", "думаю", "ясно", "также", "всей", "около", "дальше", "знаешь", "комнате", "истории", "русского", "той", "скорее", "свете", "какую", "гораздо", "легко", "брат", "имеет", "рукой", "тысяч", "разговор", "стоит", "будут", "другие", "мира", "моего", "любовь", "сих", "право", "какого", "идти", "одним", "проговорил", "часа", "своими", "разве", "чувство", "другое", "знала", "случилось", "вовсе", "руках", "мать", "двери", "комнаты", "могло", "народ", "вечер", "большой", "вполне", "могут", "лице", "случай", "письмо", "вокруг", "наш", "человеком", "понял", "голосом", "имя", "десять", "конце", "господин", "глазах", "человеку", "вероятно", "узнал", "мимо", "чрезвычайно", "хотите", "моему", "прибавил", "сильно", "говорили", "иначе", "туда", "часть", "узнать", "среди", "ваше", "вижу", "свет", "чему", "показалось", "доме", "вид", "словом", "нашего", "воды", "таком", "другого", "никак", "виде", "путь", "раньше", "найти", "сидел", "видела", "русский", "всеми", "глаз", "головой", "спокойно", "ноги", "тихо", "столько", "остановился", "минуты", "самым", "дом", "денег", "силы", "каком", "хочешь", "нею", "иметь", "давеча", "любил", "степени", "одном", "никаких", "люблю", "посмотрел", "другом", "менее", "улыбкой", "рядом", "смерти", "нечего", "думать", "самых", "никакого", "трудно", "например", "понять", "разумеется", "глядя", "народа", "вместо", "верно", "вскричал", "острове", "говорят", "возможно", "громко", "состоянии", "странно", "вошел", "молодой", "самые", "пришел", "сквозь", "четыре", "вышла", "главное", "едва", "которому", "крикнул", "мгновение", "детей", "момент", "далеко", "дать", "невольно", "весело", "первых", "двадцать", "подошел", "матери", "дней", "готов", "молча", "плато", "бывает", "говоря", "будем", "слов", "сел", "встал", "сначала", "ибо", "другую", "идет", "пожалуй", "увидел", "последний", "думала", "поэтому", "взглядом", "поняла", "вещи", "земли", "реки", "ума", "каких", "оба", "чувства", "князю", "пусть", "решительно", "пути", "тело", "стол", "удалось", "ответ", "положение", "нами", "женщины", "любит", "доктор", "берег", "шел", "всякий", "стояла", "смотреть", "никого", "год", "первого", "море", "мистер", "голове", "слышал", "нарочно", "такую", "другим", "рад", "дал", "кем", "пожалуйста", "решил", "собою", "черт", "скажу", "землю", "моряк", "возможности", "утра", "моих", "ночи", "внимание", "наши", "вечера", "части", "говорите", "дорогу", "успел", "дороге", "горы", "душе", "губы", "смысле", "нашей", "личности", "самому", "обо", "кругом", "бог", "колонистов", "права", "всему", "видели", "всяком", "шесть", "рублей", "внимательно", "улыбнулась", "мире", "приехал", "напротив", "значение", "литературы", "последнее", "очевидно", "всякого", "свобода", "меньше", "хуже", "окончательно", "видимо", "внимания", "вниз", "сознание", "раза", "сто", "человеческой", "возле", "характер", "крепко", "мужа", "дни", "идея", "следует", "виду", "кому", "неделю", "известно", "думаете", "берега", "одному", "можете", "отец", "нашел", "наверное", "воскликнул", "душу", "чувствовала", "невозможно", "вслед", "пришлось", "почувствовал", "брата", "видом", "знали", "конца", "дворца", "чрез", "второй", "вашей", "огонь", "наверно", "никакой", "пришла", "целый", "земле", "сво", "резко", "забыл", "отношении", "видимому", "движение", "прошу", "является", "тяжело", "принять", "словами", "лицом", "утром", "ждал", "насчет", "обратился", "улыбаясь", "улыбнулся", "работы", "одни", "кое", "сердца", "свободу", "угодно", "продолжала", "возможность", "обратно", "современной", "город", "рот", "часы", "вновь", "стороне", "стараясь", "медленно", "девушка", "остается", "либо", "месяц", "капитан", "судьба", "выражение", "серьезно", "смотрела", "инженера", "силой", "поздно", "окна", "культуры", "любить", "слегка", "мною", "можешь", "месяца", "платье", "вечером", "новое", "помню", "квартиру", "уверен", "ехать", "правду", "хотели", "желая", "отчего", "произошло", "богу", "господа", "вроде", "конец", "сил", "желание", "идеи", "отзывы", "груди", "общества", "состояние", "наша", "оказалось", "новый", "русских", "страница", "зная", "счастье", "представить", "ближе", "взять", "сторон", "наших", "достаточно", "глубоко", "пробормотал", "рекомендации", "раздел", "уйти", "нечто", "дети", "сознания", "философии", "концов", "классической", "популярные", "цели", "многие", "остался", "рук", "вашего", "объяснить", "человеческого", "библиотека", "авторы", "цитату", "трех", "кровь", "моим", "самую", "новинки", "бестселлеры", "боюсь", "углу", "дорогой", "отсюда", "отношения", "ваш", "поскорее", "нашли", "некоторые", "решили", "христианства", "душа", "окно", "посмотрела", "закричал", "воду", "недели", "помочь", "обе", "подле", "новой", "выше", "людьми", "благодаря", "ваши", "пристально", "рода", "получил", "страх", "руке", "взглянул", "откуда", "жаль", "пошли", "пяти", "революции", "вопросы", "стола", "которыми", "заметила", "образ", "дороги", "вышли", "лестнице", "массы", "понимал", "наше", "делу", "ребенок", "спать", "искать", "касается", "дворец", "футов", "трудом", "третий", "начало", "стояли", "ребенка", "необходимо", "ночью", "ровно", "столь", "приятно", "работу", "роде", "годы", "следующий", "жена", "делает", "отвечала", "вздор", "рассказал", "жизнью", "природы", "воздух", "власти", "другая", "дочь", "впечатление", "страшно", "увидела", "нового", "лошадь", "тысячи", "водой", "разных", "прав", "ходить", "сомнения", "последние", "зла", "дух", "улицу", "твой", "муж", "девушки", "взгляда", "отношение", "самая", "двумя", "смотря", "ждать", "постоянно", "удивлением", "происходит", "зато", "легче", "прекрасно", "заметить", "будешь", "ваша", "спросить", "миг", "грудь", "прошел", "силах", "господина", "тридцать", "отчасти", "лицу", "рассказать", "позвольте", "поднял", "леса", "смех", "хочется", "сделала", "ветер", "совести", "лес", "прошло", "человеке", "тся", "неужели", "имею", "сильнее", "осторожно", "отца", "постели", "думаешь", "пошла", "виноват", "большим", "церкви", "обращаясь", "помощь", "живет", "новые", "батюшка", "сказали", "отлично", "стены", "боль", "улице", "счастья", "заговорил", "выйти", "удовольствие", "слушал", "бросился", "общество", "никому", "рука", "голова", "принял", "собственно", "девочки", "вскоре", "вспомнил", "плохо", "неожиданно", "лежал", "ног", "удовольствием", "больно", "первой", "недавно", "видит", "произнес", "возразил", "навсегда", "разные", "некоторое", "твоей", "обществе", "слезы", "положении", "придется", "свободе", "спустя", "мыслей", "первую", "княгиня", "многое", "пора", "обратилась", "своею", "вернуться", "станет", "рано", "дам", "волосы", "полчаса", "имела", "света", "головы", "речь", "подошла", "шагов", "бывало" } };
-	public static ListHashSet<string> SHETHS1 { get; } = SHETEndinds.GetSlice(1..5).JoinIntoSingle().Filter(x => x.Length > 2).ToHashSet();
+	public static string[][] SHETEndinds { get; } = [["ото", "ост", "каз", "оло", "ово", "ако", "ени", "отор", "каза", "ста", "оро", "ест", "енн", "ала", "ств", "ольк", "аст", "нно", "елов", "огд", "тве", "ове", "рос", "овор", "льн", "ель", "тел", "век", "ова", "вал", "чен", "еск", "али", "ани", "осл", "тил", "ног", "стр", "асс", "епе", "сти", "иче", "ние", "тор", "амо", "вер", "рем", "оди", "тся", "оле", "епер", "ате", "ась", "аме", "осто", "олж", "нов", "лся", "ожн", "бра", "ума", "рил", "лас", "ольш", "ите", "ьно", "про", "изн", "рин", "рит", "оте", "ину", "ром", "няз", "ско", "ись", "ыва", "ват", "тво", "уде", "ров", "овер", "дел", "сег", "ван", "лос", "тно", "льно", "ико", "ься", "ави", "ила", "одо", "рис", "огл", "вор", "или", "ение", "лен", "нул", "рост", "ерв", "оче", "рес", "ило", "дру", "вид", "тог", "нны", "рот", "мен", "омн", "ния", "они", "ичег", "аск", "ерн", "ает", "ный", "лис", "етс", "овс", "ски", "олов", "рив", "тьс", "оне", "ибу", "ибуд", "ика", "усс", "опр", "одн", "ови", "род", "усск", "чно", "осп", "увст", "руго", "ете", "сил", "лек", "ком", "луч", "еско", "ник", "лед", "нос", "тар", "ене", "жен", "ным", "сто", "увс", "ива", "дела", "стро", "ево", "ача", "око", "олн", "нач", "аро", "ысл", "нае", "акон", "едн", "ече", "омо", "ара", "кой", "тан", "торо", "азв", "роф", "обо", "орош", "рас", "апр", "оря", "але", "ости", "ные", "ран", "авл", "рог", "ков", "дес", "иво", "опро", "тель", "або", "ина", "лав", "прос", "отр", "ться", "пят", "ное", "лась", "рист", "еме", "чес", "ход", "став", "аве", "оня", "ения", "очт", "ось", "рои", "рош", "ист", "енк", "олод", "оск", "ажд", "тви", "лекс", "лько", "ака", "озв", "авн", "едо", "лаг", "озн", "рал", "енкр", "спо", "еши", "ельн", "апи", "ооб", "мот", "твен", "олу", "енно", "рый", "иль", "вен", "рик", "оста", "озд", "очн", "менн", "рен", "вобо", "воб", "кол", "редс", "оспо", "лись", "луча", "енкроф", "шенн", "оним", "згля", "мотр", "ерев", "амет", "олож", "родо", "лаго", "иког", "стре", "стан", "ство", "роти", "олжн", "ался", "аско", "альн", "оказ", "лось", "част", "ейст", "ичес", "ейча", "олон", "омна", "тран", "илос", "аста", "венн", "ност", "орог", "абот", "евск", "ость", "сегд", "айре", "собе", "роси", "мест", "озна", "евин", "ился", "онеч", "ител", "азал", "стор", "вано", "вать", "естн", "емно", "андр", "ервы", "атьс", "руги", "енны", "вова", "роиз", "айрес", "льни", "енщи", "вите", "воег", "асск", "атер", "ыстр", "равд", "исты", "реме", "ичик", "азум", "очем", "азго", "нтер", "ечес", "овол", "ербе", "ильн", "виде", "аход", "проч", "егод", "ожал", "озмо", "ывал", "ткры", "ообщ", "енер", "рани", "олуч", "нжен", "ажет", "ател", "пиле", "есят", "одум", "елик", "рест", "твеч", "ческ", "нени", "прав", "ович", "ерво", "ерес", "азыв", "лавн", "нима", "озво", "ольн", "начи", "ходи", "алас", "евол", "ельз", "овал", "еспо", "ущес", "еств", "ердц", "брат", "лжал", "алос", "лишк", "режд", "екот", "осмо", "риня", "влен", "отча", "ризн", "пилет", "аког", "ание", "илас", "апис", "апад", "жасн", "рошл", "дног", "авер", "звес", "ерну", "етро", "рият", "ичиков", "ется", "одит", "ыход", "ерберт", "исто", "тнош", "еньг", "коль", "елиг", "аскольников", "апра", "орот", "одня", "нула", "алис", "рихо", "ачин", "льны", "кого", "поко", "итьс", "жден", "ьств", "ении", "луша", "ивал", "олно", "ажды", "обра", "обст", "олже", "ывае", "трел", "ания", "вали", "бъяс", "аетс", "инов", "вере", "юбов", "сяко", "пуст", "ично", "твер", "ален", "оман", "чита", "ннос", "рекр", "скри", "оско", "ыраж", "ронс", "едел", "елат", "исьм", "днак", "ианс", "редп", "есел", "овил", "лыша", "бщес", "ришл", "леду", "атал", "спом", "риве", "амеч", "апит", "нутр", "осте", "осси", "нием", "арак", "ловн", "ногд", "оциа", "епри", "мерт", "автр", "онят", "тара", "амог", "нако", "жида", "рави", "илип", "стью", "ричи", "вала", "ысок", "евуш", "пере", "арст", "льст", "олез", "одим", "елен", "вори", "одоб", "овар", "повн", "нить", "бращ", "траш", "редл", "альш", "осту", "есча", "остр", "снов", "остоевского", "рова", "дивл", "виже", "риго", "ораз", "риба", "еобх", "атали", "рого", "олст", "райн", "еком", "рево", "ледо", "ленн", "тави", "илис", "ован", "идим", "амер", "идет", "озвр", "варт", "евоч", "огож", "ереб", "одош", "асст", "аясь", "атур", "лыбк", "оэто", "тавл", "мысл", "аждо", "удьб", "естр", "лучи", "епре", "резв", "хват", "стве", "иозн", "тлич", "лени", "скор", "трад", "воре", "пред", "ухов", "авши", "спуг", "елан", "тпра", "дова", "ешит", "ства", "рохо", "ебед", "риех", "ержа", "лубо", "овто", "адос", "остоевский", "рудн", "ивши", "отер", "тить", "ычай", "роис", "есмо", "очув", "ерди", "держ", "ообр", "амен", "посо", "едео", "овна", "нный", "ероя", "асил", "йрто", "улся", "еобы", "ливо", "трас", "коре", "режн", "елал", "ошад", "риро", "озяй", "ульт", "асме", "яжел", "осуд", "змен", "евер", "омен", "озду", "рини", "аков", "овен", "роше", "нача", "ашег", "ными", "лыбн", "тано", "етыр", "озда", "роко", "умае", "вида", "обро", "икол", "нным", "сяки", "жени", "гром", "ится", "рово", "жнос", "анно", "ешен", "ласт", "ихин", "конч", "рове", "рика", "асте", "стат", "расн", "омне", "изав", "итат", "вает", "едеон", "вить", "акры", "исте", "оват", "ебен", "динс", "гнов", "елае", "ески", "ерье", "аспо", "оряд", "омощ", "анны", "окто", "ечен", "ойде", "дово", "олот", "льна", "риме", "диви", "орал", "тари", "овес", "чени", "спол", "ертв", "арен", "кнов", "ноги", "казы", "вадц", "ерем", "озра", "оссии", "вилс", "тказ", "нуть", "ромк", "аскр", "авал", "есьм", "стин", "знат", "ркад", "тнос", "ривы", "укам", "рони", "аним", "ругу", "ассу", "азва", "атил", "улас", "мени", "аклю", "олен", "ненн", "ложн", "одил", "ронский", "ести", "ужчи", "оряч", "дыва", "аете", "лиен", "альч", "охож", "твор", "рожа", "астасья", "наеш", "праш", "еуже", "лава", "рина", "олне", "вало", "азде", "отре", "адум", "ашем", "ассм", "бяза", "апро", "орил", "ению", "рипо", "тепа", "ский", "рикн", "окру", "анят", "одст", "вила", "ужик", "вяза", "ивот", "глая", "овре", "евоз", "стви", "оскл", "одоз", "аска", "бнос", "ятел", "астр", "ассе", "чите", "ончи", "фьев", "дцат", "асту", "авло", "бург", "твов", "иван", "ерны", "тдел", "тств", "луби", "удущ", "вопо", "юбоп", "рили", "нной", "дени", "реть", "ират", "лександрович", "данн", "слов", "ками", "азны", "лист", "вановна", "еспр", "отит", "ереж", "еетс", "етве", "итер", "есно", "едос", "одро", "ворц", "рени", "ерен", "офия", "кнул", "лексей", "адеж", "сход", "пасн", "лыба", "наль", "орфи", "быкн", "сили", "арти", "бычн", "слыш", "спыт", "спок", "полн", "трем", "омни", "олаг", "можн", "етрович", "аздр", "ослу", "еожи", "нног", "риде", "апри", "саен", "саенко", "тару", "рови", "асно", "апря", "бежд", "клон", "емей", "нные", "длеж", "едов", "ений", "аблю", "ласс", "одхо", "оспе", "ника", "звол", "ранц", "тчая", "обак", "обир", "счез", "иров", "ебол", "осад", "ораб", "янно", "вили", "уман", "веря", "ванович", "окра", "ному", "озьм", "лавя", "слав", "алек", "емен", "асче", "ающи", "нулс", "ажен", "авли", "ротя", "йртон", "аешь", "доро", "лать", "озяи", "рать", "втор", "нное", "репк", "родн", "одни", "рине", "дино", "омог", "рише", "азда", "ебедев", "урно", "деал", "бсто", "езна", "линн", "ошло", "енав", "вген", "ивая", "еньш", "чеви", "рива", "нико", "апол", "ольз", "оход", "атор", "ятно", "елом", "орма", "исти", "омещ", "верд", "ывая", "ресл", "арма", "вроп", "щени", "тепе", "еизв", "сихо", "азумихин", "робо", "ойти", "звин", "ожде", "азре", "ожен", "мина", "ысле", "озбу", "олнц", "огат", "овет", "ечег", "склю", "усты", "айте", "ыват", "ледн", "орит", "логи", "урна", "авеч", "илипповна", "очеш", "ског", "арго", "нофи", "вшис", "иват", "ажно", "ереп", "тиче", "роце", "едор", "огожин", "раси", "кажи", "огов", "азви", "вляе", "аньш", "одру", "итал", "есси", "кова", "тепан", "ченн", "альц", "тный", "овин", "шего", "ываю", "рича", "дите", "исат", "ссле", "лександровна", "елич", "няги", "шись", "роща", "квоз", "азли", "влял", "печа", "уйст", "нная", "амил", "вани", "инко", "елле", "ирок", "окры", "икну", "пуск", "елез", "ству", "адно", "ракт", "изавета", "овод", "сюша", "ступ", "оздр", "умат", "рибл", "оцел"],
+		["а", "я", "ы", "и", "е", "у", "ю", "ой", "ей", "ам", "ям", "ами", "ями", "ах", "ях", "о", "ь", "ом", "ем", "ём", "ов", "ью", "ени", "енем", "ен", "ён", "ян", "енам", "енами", "енах"],
+		["ый", "ий", "ого", "его", "ому", "ему", "ым", "им", "ая", "яя", "ую", "юю", "ою", "ею", "ое", "ее", "ые", "ие", "ых", "их", "ной", "ный", "ний", "ного", "него", "ному", "нему", "ным", "ним", "ном", "нем", "ная", "няя", "ней", "ную", "нюю", "ною", "нею", "ное", "нее", "ные", "ние", "ных", "них", "он", "на", "ня", "но", "нё", "ны", "ни"],
+		[ "ть", "ать", "еть", "ить", "оть", "уть", "ыть", "ять", "овать", "евать", "л", "ал", "ел", "ил", "ол", "ул", "ыл", "ял", "овал", "евал", "ла", "ала", "ела", "ила", "ола", "ула", "ыла", "яла", "овала", "евала", "ло", "ало", "ело", "ило", "оло", "уло", "ыло", "яло", "овало", "евало", "ли", "али", "ели", "или", "оли", "ули", "ыли", "яли", "овали", "евали", "ишь", "ит", "ите", "ешь", "ёшь", "ет", "ёт", "ете", "ёте", "аешь", "аёшь", "ает", "аёт", "аем", "аём", "аете", "аёте", "еешь", "еет", "еем", "еете", "иешь", "иёшь", "иет", "иёт", "ием", "иём", "иете", "иёте", "оешь", "оет", "оем", "оете", "уешь", "уёшь", "ует", "уёт", "уем", "уём", "уете", "уёте", "ьешь", "ьёшь", "ьет", "ьёт", "ьем", "ьём", "ьете", "ьёте", "юешь", "юёшь", "юет", "юёт", "юем", "юём", "юете", "юёте", "яешь", "яет", "яем", "яете", "ут", "ют", "ают", "еют", "иют", "оют", "уют", "ьют", "юют", "яют", "ат", "ят", "ться", "аться", "еться", "иться", "оться", "уться", "ыться", "яться", "оваться", "еваться", "лся", "ался", "елся", "ился", "олся", "улся", "ылся", "ялся", "овался", "евался", "лась", "алась", "елась", "илась", "олась", "улась", "ылась", "ялась", "овалась", "евалась", "лось", "алось", "елось", "илось", "олось", "улось", "ылось", "ялось", "овалось", "евалось", "лись", "ались", "елись", "ились", "олись", "улись", "ылись", "ялись", "овались", "евались", "усь", "юсь", "ишься", "ится", "имся", "итесь", "ешься", "ёшься", "ется", "ётся", "емся", "ёмся", "етесь", "ётесь", "аешься", "аёшься", "ается", "аётся", "аемся", "аёмся", "аетесь", "аётесь", "еешься", "еется", "еемся", "еетесь", "иется", "иётся", "иемся", "иёмся", "иетесь", "иётесь", "оешься", "оется", "оемся", "оетесь", "уешься", "уёшься", "уется", "уётся", "уемся", "уёмся", "уетесь", "уётесь", "ьешься", "ьёшься", "ьется", "ьётся", "ьемся", "ьёмся", "ьетесь", "ьётесь", "юешься", "юёшься", "юется", "юётся", "юемся", "юёмся", "юетесь", "юётесь", "яешься", "яется", "яемся", "яетесь", "утся", "ются", "аются", "еются", "иются", "оются", "уются", "ьются", "юются", "яются", "атся", "ятся", "й", "ай", "ей", "ой", "уй", "ый", "яй", "йся", "айся", "ейся", "ойся", "уйся", "ыйся", "яйся", "ись", "ься", "ущий", "ющий", "ающий", "еющий", "иющий", "оющий", "ующий", "ьющий", "юющий", "яющий", "ащий", "ящий", "ущего", "ющего", "ающего", "еющего", "иющего", "оющего", "ующего", "ьющего", "юющего", "яющего", "ащего", "ящего", "ущему", "ющему", "ающему", "еющему", "иющему", "оющему", "ующему", "ьющему", "юющему", "яющему", "ащему", "ящему", "ущим", "ющим", "ающим", "еющим", "иющим", "оющим", "ующим", "ьющим", "юющим", "яющим", "ащим", "ящим", "ущем", "ющем", "ающем", "еющем", "иющем", "оющем", "ующем", "ьющем", "юющем", "яющем", "ащем", "ящем", "ущая", "ющая", "ающая", "еющая", "иющая", "оющая", "ующая", "ьющая", "юющая", "яющая", "ащая", "ящая", "ущей", "ющей", "ающей", "еющей", "иющей", "оющей", "ующей", "ьющей", "юющей", "яющей", "ащей", "ящей", "ущую", "ющую", "ающую", "еющую", "иющую", "оющую", "ующую", "ьющую", "юющую", "яющую", "ащую", "ящую", "ущее", "ющее", "ающее", "еющее", "иющее", "оющее", "ующее", "ьющее", "юющее", "яющее", "ащее", "ящее", "ущие", "ющие", "ающие", "еющие", "иющие", "оющие", "ующие", "ьющие", "юющие", "яющие", "ащие", "ящие", "ущих", "ющих", "ающих", "еющих", "иющих", "оющих", "ующих", "ьющих", "юющих", "яющих", "ащих", "ящих", "ущими", "ющими", "ающими", "еющими", "иющими", "оющими", "ующими", "ьющими", "юющими", "яющими", "ащими", "ящими", "ущийся", "ющийся", "ающийся", "еющийся", "иющийся", "оющийся", "ующийся", "ьющийся", "юющийся", "яющийся", "ащийся", "ящийся", "ущегося", "ющегося", "ающегося", "еющегося", "иющегося", "оющегося", "ующегося", "ьющегося", "юющегося", "яющегося", "ащегося", "ящегося", "ущемуся", "ющемуся", "ающемуся", "еющемуся", "иющемуся", "оющемуся", "ующемуся", "ьющемуся", "юющемуся", "яющемуся", "ащемуся", "ящемуся", "ущимся", "ющимся", "ающимся", "еющимся", "иющимся", "оющимся", "ующимся", "ьющимся", "юющимся", "яющимся", "ащимся", "ящимся", "ущемся", "ющемся", "ающемся", "еющемся", "иющемся", "оющемся", "ующемся", "ьющемся", "юющемся", "яющемся", "ащемся", "ящемся", "ущаяся", "ющаяся", "ающаяся", "еющаяся", "иющаяся", "оющаяся", "ующаяся", "ьющаяся", "юющаяся", "яющаяся", "ащаяся", "ящаяся", "ущейся", "ющейся", "ающейся", "еющейся", "иющейся", "оющейся", "ующейся", "ьющейся", "юющейся", "яющейся", "ащейся", "ящейся", "ущуюся", "ющуюся", "ающуюся", "еющуюся", "иющуюся", "оющуюся", "ующуюся", "ьющуюся", "юющуюся", "яющуюся", "ащуюся", "ящуюся", "ущееся", "ющееся", "ающееся", "еющееся", "иющееся", "оющееся", "ующееся", "ьющееся", "юющееся", "яющееся", "ащееся", "ящееся", "ущиеся", "ющиеся", "ающиеся", "еющиеся", "иющиеся", "оющиеся", "ующиеся", "ьющиеся", "юющиеся", "яющиеся", "ащиеся", "ящиеся", "ущихся", "ющихся", "ающихся", "еющихся", "иющихся", "оющихся", "ующихся", "ьющихся", "юющихся", "яющихся", "ащихся", "ящихся", "ущимися", "ющимися", "ающимися", "еющимися", "иющимися", "оющимися", "ующимися", "ьющимися", "юющимися", "яющимися", "ащимися", "ящимися", "вший", "авший", "евший", "ивший", "овший", "увший", "ывший", "явший", "овавший", "евавший", "вшего", "авшего", "евшего", "ившего", "овшего", "увшего", "ывшего", "явшего", "овавшего", "евавшего", "вшему", "авшему", "евшему", "ившему", "овшему", "увшему", "ывшему", "явшему", "овавшему", "евавшему", "вшим", "авшим", "евшим", "ившим", "овшим", "увшим", "ывшим", "явшим", "овавшим", "евавшим", "вшем", "авшем", "евшем", "ившем", "овшем", "увшем", "ывшем", "явшем", "овавшем", "евавшем", "вшая", "авшая", "евшая", "ившая", "овшая", "увшая", "ывшая", "явшая", "овавшая", "евавшая", "вшей", "авшей", "евшей", "ившей", "овшей", "увшей", "ывшей", "явшей", "овавшей", "евавшей", "вшую", "авшую", "евшую", "ившую", "овшую", "увшую", "ывшую", "явшую", "овавшую", "евавшую", "вшее", "авшее", "евшее", "ившее", "овшее", "увшее", "ывшее", "явшее", "овавшее", "евавшее", "вшие", "авшие", "евшие", "ившие", "овшие", "увшие", "ывшие", "явшие", "овавшие", "евавшие", "вших", "авших", "евших", "ивших", "овших", "увших", "ывших", "явших", "овавших", "евавших", "вшими", "авшими", "евшими", "ившими", "овшими", "увшими", "ывшими", "явшими", "овавшими", "евавшими", "вшийся", "авшийся", "евшийся", "ившийся", "овшийся", "увшийся", "ывшийся", "явшийся", "овавшийся", "евавшийся", "вшегося", "авшегося", "евшегося", "ившегося", "овшегося", "увшегося", "ывшегося", "явшегося", "овавшегося", "евавшегося", "вшемуся", "авшемуся", "евшемуся", "ившемуся", "овшемуся", "увшемуся", "ывшемуся", "явшемуся", "овавшемуся", "евавшемуся", "вшимся", "авшимся", "евшимся", "ившимся", "овшимся", "увшимся", "ывшимся", "явшимся", "овавшимся", "евавшимся", "вшемся", "авшемся", "евшемся", "ившемся", "овшемся", "увшемся", "ывшемся", "явшемся", "овавшемся", "евавшемся", "вшаяся", "авшаяся", "евшаяся", "ившаяся", "овшаяся", "увшаяся", "ывшаяся", "явшаяся", "овавшаяся", "евавшаяся", "вшейся", "авшейся", "евшейся", "ившейся", "овшейся", "увшейся", "ывшейся", "явшейся", "овавшейся", "евавшейся", "вшуюся", "авшуюся", "евшуюся", "ившуюся", "овшуюся", "увшуюся", "ывшуюся", "явшуюся", "овавшуюся", "евавшуюся", "вшееся", "авшееся", "евшееся", "ившееся", "овшееся", "увшееся", "ывшееся", "явшееся", "овавшееся", "евавшееся", "вшиеся", "авшиеся", "евшиеся", "ившиеся", "овшиеся", "увшиеся", "ывшиеся", "явшиеся", "овавшиеся", "евавшиеся", "вшихся", "авшихся", "евшихся", "ившихся", "овшихся", "увшихся", "ывшихся", "явшихся", "овавшихся", "евавшихся", "вшимися", "авшимися", "евшимися", "ившимися", "овшимися", "увшимися", "ывшимися", "явшимися", "овавшимися", "евавшимися",
+		"имый", "емый", "аемый", "еемый", "оемый", "уемый", "юемый", "яемый", "омый", "имого", "емого", "аемого", "еемого", "оемого", "уемого", "юемого", "яемого", "омого", "имому", "емому", "аемому", "еемому", "оемому", "уемому", "юемому", "яемому", "омому", "имым", "емым", "аемым", "еемым", "оемым", "уемым", "юемым", "яемым", "омым", "имом", "емом", "аемом", "еемом", "оемом", "уемом", "юемом", "яемом", "омом", "имая", "емая", "аемая", "еемая", "оемая", "уемая", "юемая", "яемая", "омая", "имой", "емой", "аемой", "еемой", "оемой", "уемой", "юемой", "яемой", "омой", "имую", "емую", "аемую", "еемую", "оемую", "уемую", "юемую", "яемую", "омую", "имое", "емое", "аемое", "еемое", "оемое", "уемое", "юемое", "яемое", "омое", "имые", "емые", "аемые", "еемые", "оемые", "уемые", "юемые", "яемые", "омые", "имых", "емых", "аемых", "еемых", "оемых", "уемых", "юемых", "яемых", "омых", "имыми", "емыми", "аемыми", "еемыми", "оемыми", "уемыми", "юемыми", "яемыми", "омыми", "аный", "еный", "ёный", "яный", "ованый", "еваный", "ёваный", "анный", "енный", "ённый", "янный", "ованный", "еванный", "ёванный", "аного", "еного", "ёного", "яного", "ованого", "еваного", "ёваного", "анного", "енного", "ённого", "янного", "ованного", "еванного", "ёванного", "аному", "еному", "ёному", "яному", "ованому", "еваному", "ёваному", "анному", "енному", "ённому", "янному", "ованному", "еванному", "ёванному", "аным", "еным", "ёным", "яным", "ованым", "еваным", "ёваным", "анным", "енным", "ённым", "янным", "ованным", "еванным", "ёванным", "аном", "еном", "ёном", "яном", "ованом", "еваном", "ёваном", "анном", "енном", "ённом", "янном", "ованном", "еванном", "ёванном", "аная", "еная", "ёная", "яная", "ованая", "еваная", "ёваная", "анная", "енная", "ённая", "янная", "ованная", "еванная", "ёванная", "аной", "еной", "ёной", "яной", "ованой", "еваной", "ёваной", "анной", "енной", "ённой", "янной", "ованной", "еванной", "ёванной", "аную", "еную", "ёную", "яную", "ованую", "еваную", "ёваную", "анную", "енную", "ённую", "янную", "ованную", "еванную", "ёванную", "аное", "еное", "ёное", "яное", "ованое", "еваное", "ёваное", "анное", "енное", "ённое", "янное", "ованное", "еванное", "ёванное", "аные", "еные", "ёные", "яные", "ованые", "еваные", "ёваные", "анные", "енные", "ённые", "янные", "ованные", "еванные", "ёванные", "аных", "еных", "ёных", "яных", "ованых", "еваных", "ёваных", "анных", "енных", "ённых", "янных", "ованных", "еванных", "ёванных", "аными", "еными", "ёными", "яными", "оваными", "еваными", "ёваными", "анными", "енными", "ёнными", "янными", "ованными", "еванными", "ёванными", "тый", "атый", "етый", "итый", "отый", "утый", "ытый", "ятый", "того", "атого", "етого", "итого", "отого", "утого", "ытого", "ятого", "тому", "атому", "етому", "итому", "отому", "утому", "ытому", "ятому", "тым", "атым", "етым", "итым", "отым", "утым", "ытым", "ятым", "том", "атом", "етом", "итом", "отом", "утом", "ытом", "ятом", "тая", "атая", "етая", "итая", "отая", "утая", "ытая", "ятая", "той", "атой", "етой", "итой", "отой", "утой", "ытой", "ятой", "тую", "атую", "етую", "итую", "отую", "утую", "ытую", "ятую", "тое", "атое", "етое", "итое", "отое", "утое", "ытое", "ятое", "тые", "атые", "етые", "итые", "отые", "утые", "ытые", "ятые", "тых", "атых", "етых", "итых", "отых", "утых", "ытых", "ятых", "тыми", "атыми", "етыми", "итыми", "отыми", "утыми", "ытыми", "ятыми", "учи", "ючи", "аючи", "еючи", "иючи", "оючи", "уючи", "ьючи", "юючи", "яючи", "в", "ав", "ев", "ив", "ув", "ыв", "яв", "овав", "евав", "вши", "авши", "евши", "ивши", "овши", "увши", "ывши", "явши", "овавши", "евавши", "ши", "ась", "ясь", "учись", "ючись", "аючись", "еючись", "иючись", "оючись", "уючись", "ьючись", "юючись", "яючись", "вшись", "авшись", "евшись", "ившись", "овшись", "увшись", "ывшись", "явшись", "овавшись", "евавшись", "шись" ],
+		["без", "безо", "близ", "в", "вблизи", "ввиду", "вглубь", "вдогон", "вдоль", "взамен", "включая", "вкруг", "вместо", "вне", "внизу", "внутри", "внутрь", "во", "вовнутрь", "возле", "вокруг", "вопреки", "впереди", "вроде", "вслед", "вследствие", "встречу", "выключая", "для", "до", "за", "заместо", "из", "изнутри", "изо", "исключая", "к", "касаемо", "касательно", "ко", "кончая", "кроме", "кругом", "меж", "между", "мимо", "на", "наверху", "навстречу", "над", "надо", "назад", "назло", "накануне", "наперекор", "наперерез", "наподобие", "напротив", "насчёт", "ниже", "о", "об", "обо", "обок", "около", "от", "относительно", "ото", "перед", "передо", "по", "поверх", "под", "под видом", "подле", "подо", "подобно", "позади", "помимо", "поперёд", "поперёк", "порядка", "посередине", "после", "посреди", "посредине", "посредством", "пред", "предо", "прежде", "при", "про", "против", "путём", "ради", "с", "сверх", "сверху", "свыше", "сзади", "сквозь", "снизу", "со", "согласно", "спустя", "среди", "средь", "сродни", "супротив", "у", "через", "черезо", "чрез"],
+		["что", "как", "его", "это", "было", "все", "так", "она", "меня", "только", "мне", "был", "уже", "еще", "даже", "быть", "для", "ему", "сказал", "была", "когда", "очень", "они", "или", "чтобы", "есть", "него", "теперь", "себя", "если", "может", "чем", "были", "того", "время", "вас", "нибудь", "вдруг", "раз", "себе", "вот", "потому", "ничего", "этом", "этого", "нет", "человек", "вам", "будет", "сказала", "тебя", "этот", "где", "тем", "опять", "ведь", "под", "сам", "тебе", "без", "князь", "можно", "ней", "тоже", "который", "более", "ним", "тут", "всего", "больше", "человека", "всех", "жизни", "том", "мог", "нас", "потом", "там", "почти", "при", "совсем", "которые", "один", "своей", "дело", "них", "здесь", "сказать", "надо", "несколько", "эти", "тогда", "после", "свою", "глаза", "всегда", "сейчас", "какой", "кто", "день", "наконец", "знаю", "этой", "совершенно", "чего", "будто", "между", "два", "хотя", "эту", "такой", "никогда", "жизнь", "говорил", "через", "спросил", "ответил", "нам", "минуту", "нее", "хотел", "тот", "просто", "своего", "стал", "много", "над", "всем", "руку", "нужно", "другой", "слова", "руки", "перед", "хорошо", "про", "точно", "хоть", "мой", "лицо", "вместе", "именно", "людей", "свои", "эта", "лучше", "конечно", "такое", "нем", "кажется", "тому", "своим", "самом", "слишком", "которая", "тотчас", "нельзя", "три", "лет", "своих", "всю", "знал", "лишь", "чуть", "таки", "могу", "собой", "чтоб", "одно", "снова", "прежде", "этим", "голову", "которого", "нему", "одного", "стало", "весь", "свое", "почему", "давно", "свой", "которой", "хочу", "говорить", "должен", "знает", "пока", "прямо", "стороны", "деле", "мысли", "деньги", "особенно", "которых", "назад", "сама", "друг", "которое", "могла", "мысль", "люди", "этих", "мной", "моей", "довольно", "сделать", "впрочем", "одной", "колонисты", "заметил", "времени", "думал", "делать", "самого", "говорит", "одна", "быстро", "могли", "видел", "тех", "казалось", "дня", "действительно", "сегодня", "куда", "которую", "пор", "ответила", "случае", "против", "начал", "вопрос", "долго", "моя", "образом", "дверь", "оно", "какое", "ними", "сколько", "таким", "иногда", "своем", "дома", "первый", "спросила", "место", "какие", "кроме", "инженер", "часов", "немного", "буду", "знаете", "словно", "какая", "князя", "ночь", "отвечал", "такие", "свободы", "продолжал", "этому", "самое", "две", "мере", "мои", "сюда", "должно", "правда", "каким", "вся", "русской", "пять", "каждый", "сами", "слово", "часто", "острова", "сердце", "подумал", "тобой", "одну", "вообще", "однако", "двух", "любви", "хотела", "минут", "стоял", "такого", "комнату", "никто", "пред", "вами", "других", "затем", "должны", "домой", "кого", "час", "духа", "зачем", "ужасно", "говорю", "месте", "пошел", "самой", "генерал", "завтра", "вышел", "вчера", "глазами", "значит", "мало", "такая", "стала", "видеть", "голос", "должна", "дела", "говорила", "несмотря", "таких", "лица", "крайней", "жить", "весьма", "которым", "чувствовал", "хочет", "скоро", "понимаю", "года", "взгляд", "непременно", "котором", "знать", "взял", "хотелось", "имел", "сторону", "начала", "руками", "мое", "стали", "мир", "вперед", "мою", "остров", "места", "сразу", "сделал", "берегу", "равно", "видно", "женщина", "души", "самый", "друга", "своему", "смотрел", "думаю", "ясно", "также", "всей", "около", "дальше", "знаешь", "комнате", "истории", "русского", "той", "скорее", "свете", "какую", "гораздо", "легко", "брат", "имеет", "рукой", "тысяч", "разговор", "стоит", "будут", "другие", "мира", "моего", "любовь", "сих", "право", "какого", "идти", "одним", "проговорил", "часа", "своими", "разве", "чувство", "другое", "знала", "случилось", "вовсе", "руках", "мать", "двери", "комнаты", "могло", "народ", "вечер", "большой", "вполне", "могут", "лице", "случай", "письмо", "вокруг", "наш", "человеком", "понял", "голосом", "имя", "десять", "конце", "господин", "глазах", "человеку", "вероятно", "узнал", "мимо", "чрезвычайно", "хотите", "моему", "прибавил", "сильно", "говорили", "иначе", "туда", "часть", "узнать", "среди", "ваше", "вижу", "свет", "чему", "показалось", "доме", "вид", "словом", "нашего", "воды", "таком", "другого", "никак", "виде", "путь", "раньше", "найти", "сидел", "видела", "русский", "всеми", "глаз", "головой", "спокойно", "ноги", "тихо", "столько", "остановился", "минуты", "самым", "дом", "денег", "силы", "каком", "хочешь", "нею", "иметь", "давеча", "любил", "степени", "одном", "никаких", "люблю", "посмотрел", "другом", "менее", "улыбкой", "рядом", "смерти", "нечего", "думать", "самых", "никакого", "трудно", "например", "понять", "разумеется", "глядя", "народа", "вместо", "верно", "вскричал", "острове", "говорят", "возможно", "громко", "состоянии", "странно", "вошел", "молодой", "самые", "пришел", "сквозь", "четыре", "вышла", "главное", "едва", "которому", "крикнул", "мгновение", "детей", "момент", "далеко", "дать", "невольно", "весело", "первых", "двадцать", "подошел", "матери", "дней", "готов", "молча", "плато", "бывает", "говоря", "будем", "слов", "сел", "встал", "сначала", "ибо", "другую", "идет", "пожалуй", "увидел", "последний", "думала", "поэтому", "взглядом", "поняла", "вещи", "земли", "реки", "ума", "каких", "оба", "чувства", "князю", "пусть", "решительно", "пути", "тело", "стол", "удалось", "ответ", "положение", "нами", "женщины", "любит", "доктор", "берег", "шел", "всякий", "стояла", "смотреть", "никого", "год", "первого", "море", "мистер", "голове", "слышал", "нарочно", "такую", "другим", "рад", "дал", "кем", "пожалуйста", "решил", "собою", "черт", "скажу", "землю", "моряк", "возможности", "утра", "моих", "ночи", "внимание", "наши", "вечера", "части", "говорите", "дорогу", "успел", "дороге", "горы", "душе", "губы", "смысле", "нашей", "личности", "самому", "обо", "кругом", "бог", "колонистов", "права", "всему", "видели", "всяком", "шесть", "рублей", "внимательно", "улыбнулась", "мире", "приехал", "напротив", "значение", "литературы", "последнее", "очевидно", "всякого", "свобода", "меньше", "хуже", "окончательно", "видимо", "внимания", "вниз", "сознание", "раза", "сто", "человеческой", "возле", "характер", "крепко", "мужа", "дни", "идея", "следует", "виду", "кому", "неделю", "известно", "думаете", "берега", "одному", "можете", "отец", "нашел", "наверное", "воскликнул", "душу", "чувствовала", "невозможно", "вслед", "пришлось", "почувствовал", "брата", "видом", "знали", "конца", "дворца", "чрез", "второй", "вашей", "огонь", "наверно", "никакой", "пришла", "целый", "земле", "сво", "резко", "забыл", "отношении", "видимому", "движение", "прошу", "является", "тяжело", "принять", "словами", "лицом", "утром", "ждал", "насчет", "обратился", "улыбаясь", "улыбнулся", "работы", "одни", "кое", "сердца", "свободу", "угодно", "продолжала", "возможность", "обратно", "современной", "город", "рот", "часы", "вновь", "стороне", "стараясь", "медленно", "девушка", "остается", "либо", "месяц", "капитан", "судьба", "выражение", "серьезно", "смотрела", "инженера", "силой", "поздно", "окна", "культуры", "любить", "слегка", "мною", "можешь", "месяца", "платье", "вечером", "новое", "помню", "квартиру", "уверен", "ехать", "правду", "хотели", "желая", "отчего", "произошло", "богу", "господа", "вроде", "конец", "сил", "желание", "идеи", "отзывы", "груди", "общества", "состояние", "наша", "оказалось", "новый", "русских", "страница", "зная", "счастье", "представить", "ближе", "взять", "сторон", "наших", "достаточно", "глубоко", "пробормотал", "рекомендации", "раздел", "уйти", "нечто", "дети", "сознания", "философии", "концов", "классической", "популярные", "цели", "многие", "остался", "рук", "вашего", "объяснить", "человеческого", "библиотека", "авторы", "цитату", "трех", "кровь", "моим", "самую", "новинки", "бестселлеры", "боюсь", "углу", "дорогой", "отсюда", "отношения", "ваш", "поскорее", "нашли", "некоторые", "решили", "христианства", "душа", "окно", "посмотрела", "закричал", "воду", "недели", "помочь", "обе", "подле", "новой", "выше", "людьми", "благодаря", "ваши", "пристально", "рода", "получил", "страх", "руке", "взглянул", "откуда", "жаль", "пошли", "пяти", "революции", "вопросы", "стола", "которыми", "заметила", "образ", "дороги", "вышли", "лестнице", "массы", "понимал", "наше", "делу", "ребенок", "спать", "искать", "касается", "дворец", "футов", "трудом", "третий", "начало", "стояли", "ребенка", "необходимо", "ночью", "ровно", "столь", "приятно", "работу", "роде", "годы", "следующий", "жена", "делает", "отвечала", "вздор", "рассказал", "жизнью", "природы", "воздух", "власти", "другая", "дочь", "впечатление", "страшно", "увидела", "нового", "лошадь", "тысячи", "водой", "разных", "прав", "ходить", "сомнения", "последние", "зла", "дух", "улицу", "твой", "муж", "девушки", "взгляда", "отношение", "самая", "двумя", "смотря", "ждать", "постоянно", "удивлением", "происходит", "зато", "легче", "прекрасно", "заметить", "будешь", "ваша", "спросить", "миг", "грудь", "прошел", "силах", "господина", "тридцать", "отчасти", "лицу", "рассказать", "позвольте", "поднял", "леса", "смех", "хочется", "сделала", "ветер", "совести", "лес", "прошло", "человеке", "тся", "неужели", "имею", "сильнее", "осторожно", "отца", "постели", "думаешь", "пошла", "виноват", "большим", "церкви", "обращаясь", "помощь", "живет", "новые", "батюшка", "сказали", "отлично", "стены", "боль", "улице", "счастья", "заговорил", "выйти", "удовольствие", "слушал", "бросился", "общество", "никому", "рука", "голова", "принял", "собственно", "девочки", "вскоре", "вспомнил", "плохо", "неожиданно", "лежал", "ног", "удовольствием", "больно", "первой", "недавно", "видит", "произнес", "возразил", "навсегда", "разные", "некоторое", "твоей", "обществе", "слезы", "положении", "придется", "свободе", "спустя", "мыслей", "первую", "княгиня", "многое", "пора", "обратилась", "своею", "вернуться", "станет", "рано", "дам", "волосы", "полчаса", "имела", "света", "головы", "речь", "подошла", "шагов", "бывало"]];
+	public static ListHashSet<string> SHETHS1 { get; } = [.. SHETEndinds.GetSlice(1..5).JoinIntoSingle().Filter(x => x.Length > 2)];
 	public static int SHETThreshold1 { get; } = ValuesInByte - 2 - (SHETHS1.Length == ValuesInByte - 2 ? 0 : Max(SHETHS1.Length, 0) / (ValuesInByte - 2));
-	public static ListHashSet<string> SHETHS2 { get; } = SHETEndinds.GetSlice(4..).Reverse().JoinIntoSingle().Filter(x => x.Length > 2).ToHashSet();
+	public static ListHashSet<string> SHETHS2 { get; } = [.. SHETEndinds.GetSlice(4..).Reverse().JoinIntoSingle().Filter(x => x.Length > 2)];
 	public static int SHETThreshold2 { get; } = ValuesInByte - 2 - (SHETHS2.Length == ValuesInByte - 2 ? 0 : Max(SHETHS2.Length, 0) / (ValuesInByte - 2));
 	public static Encoding Encoding1251 { get; } = Encoding.GetEncoding(1251);
 }
@@ -78,7 +80,7 @@ public class Decoding : AresTLib007.Decoding
 	public override byte[] Decode(byte[] compressedFile, byte encodingVersion)
 	{
 		if (compressedFile.Length <= 2)
-			return Array.Empty<byte>();
+			return [];
 		if (encodingVersion == 0)
 			return compressedFile;
 		else if (encodingVersion < ProgramVersion)
@@ -97,7 +99,7 @@ public class Decoding : AresTLib007.Decoding
 		Current[0] += ProgressBarStep;
 		if (rle == 7)
 			byteList = byteList.DecodeRLE();
-		return byteList.Repeat(repeatsCount).ToArray();
+		return [.. byteList.Repeat(repeatsCount)];
 	}
 
 	protected override void SplitMethod(int method)
@@ -110,7 +112,7 @@ public class Decoding : AresTLib007.Decoding
 	{
 		if (utf8)
 		{
-			list.Add(new() { new() { new(ar.ReadEqual(2), 2) } });
+			list.Add([[new(ar.ReadEqual(2), 2)]]);
 			if (list[^1][0][0].Lower == 1)
 				list[^1][0].Add(new(ar.ReadEqual(ValuesInByte), ValuesInByte));
 		}
@@ -138,7 +140,7 @@ public class Decoding : AresTLib007.Decoding
 	{
 		Status[0] = 0;
 		StatusMaximum[0] = counter;
-		List<ShortIntervalList> result = new();
+		List<ShortIntervalList> result = [];
 		var startingArithmeticMap = lz == 0 ? huffmanData.ArithmeticMap : huffmanData.ArithmeticMap[..^1];
 		var uniqueLists = spaceCodes ? RedStarLinq.Fill(2, i => huffmanData.UniqueList.PConvert(x => new ShortIntervalList { x, new((uint)i, 2) })) : huffmanData.UniqueList.Convert(x => new ShortIntervalList() { x });
 		for (; counter > 0; counter--, Status[0]++)
@@ -208,7 +210,7 @@ public class Decoding : AresTLib007.Decoding
 		Status[0] = 0;
 		StatusMaximum[0] = GetArrayLength(input.Length, BWTBlockSize + BWTBlockExtraSize);
 		var bytes = input.NConvert(x => (byte)x[0].Lower);
-		NList<byte> bytes2 = new();
+		NList<byte> bytes2 = [];
 		for (var i = 0; i < bytes.Length;)
 		{
 			var rle = bytes[i] & ValuesInByte >> 1;
@@ -247,14 +249,14 @@ public class Decoding : AresTLib007.Decoding
 		for (var i = 0; i < input.Length; i++)
 		{
 			it = convert[it];
-			result[i] = new() { new(input[it], ValuesInByte) };
+			result[i] = [new(input[it], ValuesInByte)];
 		}
 		return result;
 	}
 
 	public override NList<byte> DecodeRLEAfterBWT(NList<byte> byteList, ref int i)
 	{
-		NList<byte> result = new();
+		NList<byte> result = [];
 		int length, serie, l;
 		byte temp;
 		for (; i < byteList.Length && result.Length < BWTBlockSize;)
@@ -319,7 +321,7 @@ public class Decoding : AresTLib007.Decoding
 
 	protected virtual string DecodeSHET2(string text)
 	{
-		String chars = new(), tempChars = new();
+		String chars = [], tempChars = [];
 		var state = 0;
 		for (var i = 0; i < text.Length; i++)
 		{
