@@ -1,2 +1,2 @@
 ﻿// See https://aka.ms/new-console-template for more information
-AresTLib.MainClass.Main(new[] { "11000" });
+AresTLib.MainClass.Main(["11000"]);
